Refactor extractors
